% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{.commonDetailEndpoint}
\alias{.commonDetailEndpoint}
\title{Handle common detail endpoints}
\usage{
.commonDetailEndpoint(type, id)
}
\arguments{
\item{type}{Type of detail endpoint}

\item{id}{The internal CIViC ID}

\item{count}{the number of assertions to retrieve}
}
\description{
Handle common detail endpoints
}
