% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{getAllVariantGroups}
\alias{getAllVariantGroups}
\title{Get a list of variant groups}
\usage{
getAllVariantGroups(page = 1, count = 25)
}
\arguments{
\item{page}{the page number to retrieve}

\item{count}{the number of variant groups to retrieve}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve all variant groups from the CIViC DB
}
\examples{
getAllVariantGroups(count = 10)
}
\keyword{groups}
\keyword{variant}
