% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{getAllVariants}
\alias{getAllVariants}
\title{Get a list of variants}
\usage{
getAllVariants(page = 1, count = 25)
}
\arguments{
\item{page}{the page number to retrieve}

\item{count}{the number of variants to retrieve}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve all variants from the CIViC DB
}
\examples{
getAllVariants(count = 10)
getAllVariants(page = 2, count = 10)
}
\keyword{variants}
