% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{getAssertion}
\alias{getAssertion}
\title{Get a specific assertion}
\usage{
getAssertion(id)
}
\arguments{
\item{id}{Internal CIViC ID of the assertion of interest}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve a specific assertion from the CIViC DB
}
\examples{
getAssertion(id = 1)
}
\keyword{assertion}
