% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{getAssertionMetadata}
\alias{getAssertionMetadata}
\title{Get a specific assertion metadata information}
\usage{
getAssertionMetadata(id, type)
}
\arguments{
\item{id}{Internal CIViC ID of the assertion of interest}

\item{type}{Type of metadata (comments, suggested_changes, revisions)}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve metadata information for a specific variant group from the CIViC DB
}
\examples{
getAssertionMetadata(id = 1, type = "comments")
}
\keyword{assertion,}
\keyword{metadata}
