% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{getEvidenceItemMetadata}
\alias{getEvidenceItemMetadata}
\title{Get a specific evidence item metadata information}
\usage{
getEvidenceItemMetadata(id, type)
}
\arguments{
\item{id}{Internal CIViC ID of the evidence item of interest}

\item{type}{Type of metadata (comments, suggested_changes, revisions)}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve metadata information for a specific evidence item from the CIViC DB
}
\examples{
getEvidenceItemMetadata(id = 1, type = "comments")
}
\keyword{evidence}
\keyword{item,}
\keyword{metadata}
