% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{getGene}
\alias{getGene}
\title{Get a specific gene}
\usage{
getGene(id, identifier_type = "civic_id")
}
\arguments{
\item{id}{ID of the gene of interest}

\item{identifier_type}{Type of gene identifier (entrez_id, entrez_symbol, civic_id)}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve a specific gene from the CIViC DB
}
\examples{
getGene(id = 1)
getGene(id = "ALK", identifier_type = "entrez_symbol")
getGene(id = 238, identifier_type = "entrez_id")
}
\keyword{gene}
