% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{getGeneMetadata}
\alias{getGeneMetadata}
\title{Get a specific gene metadata}
\usage{
getGeneMetadata(id, type)
}
\arguments{
\item{id}{ID of the gene of interest}

\item{type}{Type of metadata (comments, suggested_changes, revisions, variants)}

\item{identifier_type}{Type of gene identifier (entrez_id, entrez_symbol, civic_id)}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve metadata for a specific gene from the CIViC DB
}
\examples{
getGeneMetadata(id = 1, type = "comments")
getGeneMetadata(id = 1, type = "variants")
}
\keyword{gene,}
\keyword{metadata}
