% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civicApi.R
\name{getVariantMetadata}
\alias{getVariantMetadata}
\title{Get a specific variant metadata information}
\usage{
getVariantMetadata(id, type)
}
\arguments{
\item{id}{Internal CIViC ID of the variant of interest}

\item{type}{Type of metadata (comments, suggested_changes, revisions, evidence_items)}
}
\value{
An S3 Object of type civic_api containing the content, url, and response
}
\description{
Retrieve metadata information for a specific variant from the CIViC DB
}
\examples{
getVariantMetadata(id = 1, type = "comments")
getVariantMetadata(id = 1, type = "revisions")
}
\keyword{metadata}
\keyword{variant,}
